+++ crates/warpgate-pdk/src/lib.rs
@@ -1,11 +1,15 @@
+//! Reusable WASM macros and functions for plugin developer kits (PDKs).
+
 mod api;
 mod funcs;
 mod macros;
+mod path_ext;
 #[cfg(feature = "tracing")]
 mod tracing;
 
 pub use api::*;
 pub use funcs::*;
+pub use path_ext::*;
 #[cfg(feature = "tracing")]
 pub use tracing::*;
 pub use warpgate_api::*;