+++ crates/warpgate/src/host.rs
@@ -1,5 +1,7 @@
+//! Host functions that are executed on the host machine, and are exposed
+//! to WASM plugins (the guest) through the PDKs.
+
 use crate::clients::{HttpClient, WarpgateHttpClientError};
-use crate::helpers;
 use crate::plugin_error::WarpgatePluginError;
 use extism::{CurrentPlugin, Error, Function, UserData, Val, ValType};
 use starbase_shell::{ShellType, join_exe_args};
@@ -16,15 +18,22 @@ use tokio::runtime::Handle;
 use tracing::{debug, error, instrument, trace, warn};
 use warpgate_api::{
     ExecCommandInput, ExecCommandOutput, HostLogInput, HostLogTarget, SendRequestInput,
-    SendRequestOutput,
+    SendRequestOutput, convert_to_real_native_path,
 };
 
 /// Data passed to each host function.
 #[derive(Clone)]
 pub struct HostData {
+    /// Location where cached files are stored.
     pub cache_dir: PathBuf,
+
+    /// Instance of our HTTP client, used for sending requests.
     pub http_client: Arc<HttpClient>,
+
+    /// Mapping of virtual paths, from host to guest paths.
     pub virtual_paths: Vec<(PathBuf, PathBuf)>,
+
+    /// Current working directory, in which commands are executed.
     pub working_dir: PathBuf,
 }
 
@@ -37,13 +46,6 @@ pub fn create_host_functions(data: HostData) -> Vec<Function> {
             [ValType::I64],
             UserData::new(data.clone()),
             exec_command,
-        ),
-        Function::new(
-            "from_virtual_path",
-            [ValType::I64],
-            [ValType::I64],
-            UserData::new(data.clone()),
-            from_virtual_path,
         ),
         Function::new(
             "get_env_var",
@@ -66,13 +68,6 @@ pub fn create_host_functions(data: HostData) -> Vec<Function> {
             [],
             UserData::new(data.clone()),
             set_env_var,
-        ),
-        Function::new(
-            "to_virtual_path",
-            [ValType::I64],
-            [ValType::I64],
-            UserData::new(data.clone()),
-            to_virtual_path,
         ),
     ]
 }
@@ -160,7 +155,7 @@ fn exec_command(
 
     // Relative or absolute file path
     let maybe_exe = if input.command.contains('/') || input.command.contains('\\') {
-        let path = helpers::from_virtual_path(&data.virtual_paths, PathBuf::from(&input.command));
+        let path = convert_to_real_native_path(&input.command, &data.virtual_paths);
 
         if path.exists() {
             // This is temporary since WASI does not support updating file permissions yet!
@@ -186,11 +181,11 @@ fn exec_command(
     };
 
     // Determine working directory
-    let cwd = if let Some(cwd) = &input.cwd {
-        helpers::from_virtual_path(&data.virtual_paths, cwd)
-    } else {
-        data.working_dir.clone()
-    };
+    let cwd = input
+        .cwd
+        .as_ref()
+        .map(|cwd| convert_to_real_native_path(cwd, &data.virtual_paths))
+        .unwrap_or_else(|| data.working_dir.clone());
 
     // Determine the shell
     let shell_name = input.shell.or_else(|| env::var("PROTO_SHELL").ok());
@@ -230,7 +225,7 @@ fn exec_command(
             input
                 .paths
                 .iter()
-                .map(|virtual_path| helpers::from_virtual_path(&data.virtual_paths, virtual_path)),
+                .map(|path| convert_to_real_native_path(path, &data.virtual_paths)),
         );
         paths.extend(env_paths);
 
@@ -449,7 +444,7 @@ fn set_env_var(
         let new_path = value
             .replace(';', ":")
             .split(':')
-            .map(|path| helpers::from_virtual_path(&data.virtual_paths, PathBuf::from(path)))
+            .map(|path| convert_to_real_native_path(path, &data.virtual_paths))
             .collect::<Vec<_>>();
 
         trace!(
@@ -475,72 +470,6 @@ fn set_env_var(
 
         unsafe { env::set_var(name, value) };
     }
-
-    Ok(())
-}
-
-#[instrument(name = "host_func_from_virtual_path", skip_all)]
-fn from_virtual_path(
-    plugin: &mut CurrentPlugin,
-    inputs: &[Val],
-    outputs: &mut [Val],
-    user_data: UserData<HostData>,
-) -> Result<(), Error> {
-    let original_path = PathBuf::from(plugin.memory_get_val::<String>(&inputs[0])?);
-    let uuid = plugin.id().to_string();
-
-    trace!(
-        plugin = &uuid,
-        original_path = ?original_path,
-        "Calling host function {}",
-        color::label("from_virtual_path"),
-    );
-
-    let data = user_data.get()?;
-    let data = data.lock().unwrap();
-    let real_path = helpers::from_virtual_path(&data.virtual_paths, &original_path);
-
-    trace!(
-        plugin = &uuid,
-        real_path = ?real_path,
-        "Called host function {}",
-        color::label("from_virtual_path"),
-    );
-
-    plugin.memory_set_val(&mut outputs[0], real_path.to_string_lossy().to_string())?;
-
-    Ok(())
-}
-
-#[instrument(name = "host_func_to_virtual_path", skip_all)]
-fn to_virtual_path(
-    plugin: &mut CurrentPlugin,
-    inputs: &[Val],
-    outputs: &mut [Val],
-    user_data: UserData<HostData>,
-) -> Result<(), Error> {
-    let original_path = PathBuf::from(plugin.memory_get_val::<String>(&inputs[0])?);
-    let uuid = plugin.id().to_string();
-
-    trace!(
-        plugin = &uuid,
-        original_path = ?original_path,
-        "Calling host function {}",
-        color::label("to_virtual_path"),
-    );
-
-    let data = user_data.get()?;
-    let data = data.lock().unwrap();
-    let virtual_path = helpers::to_virtual_path(&data.virtual_paths, &original_path);
-
-    trace!(
-        plugin = &uuid,
-        virtual_path = ?virtual_path.virtual_path(),
-        "Called host function {}",
-        color::label("to_virtual_path"),
-    );
-
-    plugin.memory_set_val(&mut outputs[0], serde_json::to_string(&virtual_path)?)?;
 
     Ok(())
 }