+++ CHANGELOG.md
@@ -15,6 +15,28 @@
 - [Rust](https://github.com/moonrepo/plugins/blob/master/tools/rust/CHANGELOG.md)
 - [Schema (TOML, JSON, YAML)](https://github.com/moonrepo/plugins/blob/master/tools/internal-schema/CHANGELOG.md)
 
+## Unreleased
+
+#### 💥 Breaking
+
+- **WASM API**
+  - Reworked the `VirtualPath` type from the ground up. Is no longer an enum, but instead a newtype wrapper around `PathBuf`.
+    - This allows for better interoperability with the Rust ecosystem, and makes it easier to work with virtual paths in general.
+    - Additionally, because of this change, `VirtualPath` now has access to all `PathBuf` methods, which was not possible before.
+  - Removed the `from_virtual_path` and `to_virtual_path` extism host functions, `into_real_path` and `into_virtual_path` wrapper functions, and `real_path!` and `virtual_path!` macros. Use the conversion utils instead (below).
+  - Removed `PluginUnresolvedContext.tool_dir` and `version` fields.
+  - Removed `LocateExecutablesOutput.exes_dir` field, use `exes_dirs` instead.
+  - Updated `CommandInstruction.cwd` to be a `VirtualPath`.
+  - Updated `ExecCommandInput.paths` to be a `PathBuf`.
+
+#### 🚀 Updates
+
+- Updated plugin function calls to run on a separate blocking thread, instead of blocking the main thread.
+- **WASM API**
+  - Added a `RealPath` type, which is a newtype wrapper around `PathBuf` that represents a real path on the host file system. This is a sibling to the `VirtualPath` type, which represents a virtual path in the guest WASM environment.
+  - Added `convert_to_virtual_path` and `convert_to_real_path` helper functions for converting between real and virtual paths, using a list of host-to-guest path mappings.
+    - Can also use `VirtualPathExt::to_real_path` and `RealPathExt::to_virtual_path` extension traits for the same functionality.
+
 ## 0.59.0
 
 #### 💥 Breaking